python/sentiment-batch/pyproject.toml
[tool.poetry]
name = "bsky-sentiment-batch"
version = "0.1.0"
description = ""
authors = ["Jaz <ericvolp12@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.10"
pydantic = "^1.10.7"
transformers = "^4.29.0"
numpy = "^1.24.3"
scipy = "^1.10.1"
psycopg2-binary = "^2.9.6"
SQLAlchemy = "^2.0.13"
torch = "^2.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
